added comments and some renaming
